Move boardsProduct and boardsApp out of mattermost-plugin directory.
