+++ server/product/api_adapter.go
